fixed deserialization TLChatFull model changed chatId data type to long

+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TelegramApi.java
@@ -450,7 +450,7 @@ public interface TelegramApi {
 
     TLAbsUpdates messagesStartBot(TLAbsInputUser bot, TLAbsInputPeer peer, long randomId, String startParam) throws RpcErrorException, IOException;
 
-    TLAbsUpdates messagesToggleChatAdmins(int chatId, boolean enabled) throws RpcErrorException, IOException;
+    TLAbsUpdates messagesToggleChatAdmins(long chatId, boolean enabled) throws RpcErrorException, IOException;
 
     TLBool messagesToggleDialogPin(boolean pinned, TLAbsInputPeer peer) throws RpcErrorException, IOException;
